+++ barshelf/submissions/time_limit_exceeded/brute-force.c
@@ -9,13 +9,13 @@ int main(int argv, char** args)
 	int* arr = (int*) malloc(n * sizeof(int));
 	for (int i = 0; i < n; ++i)
 		scanf("%d", &arr[i]);
-	int res = 0;
+	long long res = 0;
 	for (int i = 0; i < n - 1; ++i)
 		for (int j = i + 1; j < n; ++j)
 			if (arr[i] > 2*arr[j])
 				++res;		
 	
-	printf("%d", res);
+	printf("%lld", res);
 	return 0; 
 }
